+++ CDBTo3DTiles/include/CDBTo3DTiles.h
@@ -2,6 +2,8 @@
 
 #include <filesystem>
 #include <memory>
+#include <string>
+#include <vector>
 
 namespace CDBTo3DTiles {
 class GlobalInitializer
@@ -18,6 +20,8 @@ public:
     Converter(const std::filesystem::path &CDBPath, const std::filesystem::path &outputPath);
 
     ~Converter() noexcept;
+
+    void combineDataset(const std::vector<std::string> &datasets);
 
     void setGenerateElevationNormal(bool generateElevationNormal);