+++ Core/src/BoundingRegion.cpp
@@ -1,4 +1,5 @@
 #include "BoundingRegion.h"
+#include "glm/glm.hpp"
 
 namespace Core {
 
@@ -8,4 +9,10 @@ BoundingRegion::BoundingRegion(const GlobeRectangle &rectangle, double minimumHe
     , m_maximumHeight{maximumHeight}
 {}
 
+BoundingRegion BoundingRegion::computeUnion(const BoundingRegion &other) const
+{
+    return BoundingRegion(m_rectangle.computeUnion(other.m_rectangle),
+                          glm::min(m_minimumHeight, other.m_minimumHeight),
+                          glm::max(m_maximumHeight, other.m_maximumHeight));
+}
 } // namespace Core