+++ Tests/CDBElevationTest.cpp
@@ -360,7 +360,7 @@ TEST_CASE("Test conversion when elevation has more LOD than imagery", "[CDBEleva
         checkAllConvertedImagery(input / "Tiles" / "N32" / "W118" / "004_Imagery", textureOutputDir, 10);
 
         // verified tileset.json
-        std::filesystem::path tilesetPath = elevationOutputDir / "tileset.json";
+        std::filesystem::path tilesetPath = elevationOutputDir / "N32W118_D001_S001_T001.json";
         REQUIRE(std::filesystem::exists(tilesetPath));
 
         std::ifstream verifiedJS(input / "VerifiedTileset.json");
@@ -390,7 +390,7 @@ TEST_CASE("Test conversion when elevation has more LOD than imagery", "[CDBEleva
         checkAllConvertedImagery(input / "Tiles" / "N32" / "W118" / "004_Imagery", textureOutputDir, 13);
 
         // verified tileset.json
-        std::filesystem::path tilesetPath = elevationOutputDir / "tileset.json";
+        std::filesystem::path tilesetPath = elevationOutputDir / "N32W118_D001_S001_T001.json";
         REQUIRE(std::filesystem::exists(tilesetPath));
 
         std::ifstream verifiedJS(input / "VerifiedTileset.json");
@@ -426,7 +426,7 @@ TEST_CASE("Test conversion when imagery has more LOD than elevation", "[CDBEleva
         checkElevationDuplicated(imageryInput, elevationOutputDir, 13);
 
         // verified tileset.json
-        std::filesystem::path tilesetPath = elevationOutputDir / "tileset.json";
+        std::filesystem::path tilesetPath = elevationOutputDir / "N32W118_D001_S001_T001.json";
         REQUIRE(std::filesystem::exists(tilesetPath));
 
         std::ifstream verifiedJS(input / "VerifiedTileset.json");
@@ -459,7 +459,7 @@ TEST_CASE("Test conversion when imagery has more LOD than elevation", "[CDBEleva
         checkElevationDuplicated(imageryInput, elevationOutputDir, 18);
 
         // verified tileset.json
-        std::filesystem::path tilesetPath = elevationOutputDir / "tileset.json";
+        std::filesystem::path tilesetPath = elevationOutputDir / "N32W118_D001_S001_T001.json";
         REQUIRE(std::filesystem::exists(tilesetPath));
 
         std::ifstream verifiedJS(input / "VerifiedTileset.json");