+++ CHANGES.md
@@ -4,6 +4,7 @@ Change Log
 ### 0.0.1 - 2020-??-??
 
 * Fixed a bug where GS model attributes and positions couldn't be processed due to too many files being open. [#23](https://github.com/CesiumGS/cdb-to-3dtiles/pull/23)
+* Provide `--combine` option to combine multiple tilesets into one. [#19](https://github.com/CesiumGS/cdb-to-3dtiles/issues/19)
 * Fixed a bug where empty simplified terrain mesh is exported to gltf. [#25](https://github.com/CesiumGS/cdb-to-3dtiles/pull/25)
 * Fixed a bug where leaf tiles were being given non-zero geometric errors. [#36](https://github.com/CesiumGS/cdb-to-3dtiles/pull/36)